Test space, basic movement, script folder, material folder, test scene, temp character.
Made a test scene with a testing space within it. Added basic left and right movement with the A & D keys, as well as the Left and Right arrow keys. These can be removed in future iterations. Script folder has been added, this folder will be reserved for MAJOR SCRIPTS ONLY, any small or personal scripts should be done in a personal script folder to prevent conflicts in later commits. Added material folder for organization, as well as a temp character material. Added a temp character model, Empty Game Object called Character parents a Cube with a 2D Box Collider. Character script is on the cube, as well as a Rigidbody2D.