New Movement, and Added Jump
Added two jump scripts, one is a basic, very floaty jump script, the 2nd adds a more drastic gravity to the end of the jump. New movement using forces allowing for acceleration. Feels very much like an ice level, not the best movement but it's a start.

Known bugs: Jump can be continuously pressed.
If the "JumpFloaty" and "Jump Improvement script are on the character, the "Character Movement" script won't function.